requirements.txt
chromadb>=1.1.0
fastmcp>=2.12.0
fastapi>=0.119.0
uvicorn>=0.37.0
sentence-transformers>=5.1.0
torch>=2.8.0
numpy>=2.3.0
openai>=2.3.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
jinja2>=3.1.0
